pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "meta-ads-cli"
version = "0.1.0"
description = "Create and manage Meta (Facebook/Instagram) ad campaigns from your terminal."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Attainment Labs", email = "hello@attainment.co" },
]
keywords = ["meta", "facebook", "instagram", "ads", "cli", "marketing", "advertising"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet",
    "Topic :: Office/Business",
]
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.scripts]
meta-ads = "meta_ads.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["meta_ads"]

[project.urls]
Homepage = "https://github.com/attainmentlabs/meta-ads-cli"
Issues = "https://github.com/attainmentlabs/meta-ads-cli/issues"
